Organize Deprecated Code
