+++ .golangci.yaml
@@ -134,11 +134,11 @@ issues:
       linters:
         - staticcheck
       text: SA1019
-    - path: pkg/client/live/legacy.go
+    - path: pkg/client/live/deprecated.go
       linters:
         - staticcheck
       text: SA1019
-    - path: pkg/api/prerecorded/v1/legacy.go
+    - path: pkg/api/prerecorded/v1/deprecated.go
       linters:
         - staticcheck
       text: SA1019
