+++ pkg/client/speak/client.go
@@ -227,37 +227,3 @@ Notes:
 func NewWSUsingChanWithCancel(ctx context.Context, ctxCancel context.CancelFunc, apiKey string, cOptions *interfaces.ClientOptions, sOptions *interfaces.WSSpeakOptions, callback msginterfaces.SpeakMessageChan) (*speakv1ws.WSChannel, error) {
 	return speakv1ws.NewUsingChanWithCancel(ctx, ctxCancel, apiKey, cOptions, sOptions, callback)
 }
-
-/***********************************/
-// Deprecated (THESE WILL STILL WORK,
-// BUT WILL BE REMOVED IN A FUTURE RELEASE)
-/***********************************/
-/*
-Legacy Client Name
-
-Deprecated: This struct is deprecated. Please use RestClient struct. This will be removed in a future release.
-*/
-type Client = speakv1rest.RESTClient
-
-/*
-NewWithDefaults creates a new speak client with all default options
-
-Deprecated: This function is deprecated. Please use NewREST(). This will be removed in a future release.
-*/
-func NewWithDefaults() *speakv1rest.RESTClient {
-	return speakv1rest.NewWithDefaults()
-}
-
-/*
-New creates a new speak client with the specified options
-
-Input parameters:
-- ctx: context.Context object
-- apiKey: string containing the Deepgram API key
-- options: ClientOptions which allows overriding things like hostname, version of the API, etc.
-
-Deprecated: This function is deprecated. Please use NewREST(). This will be removed in a future release.
-*/
-func New(apiKey string, options *interfaces.ClientOptions) *speakv1rest.RESTClient {
-	return speakv1rest.New(apiKey, options)
-}
